Customerscript

+++ Customer/Gemini-Chef.sh
@@ -11,10 +11,13 @@ then
 #	echo "Try Running Again with Corrrect Login Credentials. ERROR: Login Failed...Exiting..."
 #	exit
 #  fi         
-  echo "Pull Chef Server from Internal Registry..."
+  echo "Pull Chef Server from Registry..."
   docker pull registry.gemini-systems.net/gemini/gemini-chef
   echo "Continue to run chef ..."
-  docker rm -f gemini-chef
+  if docker ps -a |grep -aq gemini-chef; then
+        docker rm -f gemini-chef
+  fi
+
   ip=`curl -s http://ipecho.net/plain; echo`
   docker run -it -p 443:443 --privileged -v /etc/chef-server/ --name gemini-chef -h $ip -d registry.gemini-systems.net/gemini/gemini-chef
 fi
